+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - cohttp-eio: Don't blow up `Server.callback` on client disconnections. (mefyl #1015)
+- http: Fix assertion in `Source.to_string_trim` when `pos <> 0` (mefyl #1017)
 
 ## v6.0.0~beta2 (2024-01-05)